v.2.24.0505
Version notes :
- added toggle switch for turning ON/OFF visibility of RE expenses and incomes
- properties are now automatically created as soon as RE is created.
- After manager complete properties setup auto fees are automatically applied for every property in the RE. This will allow manager to finish parametrization
for all properties in the beginning of the RE no matter if property has assigned owner or not. The platform is now focused on servicing to manager, instead of owners first.
- Every owner now instead of registering his property has to make a request for registration. If manager approves obtaining of the current property by this user(owner)
the property ownership is updated. This will give access to new owner to property stored data.
- Manager can now remove owner of property without deleting the whole property. Only the ownership date is updated.
- Owner can now add or remove a property (instead of owner as in previous versions of the platform).
The control of creating and deleting properties is now transffered from user to manager.
- Some server side performance improvements were made